LayoutTests/platform/chromium/fast/css/fontloader-multiple-families-expected.txt

Test load events for fontloader.loadFont() with multiple font families.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS events['loading'] is 1
PASS events['loadstart'] is 2
PASS events['load'] is 2
PASS events['error'] is undefined
PASS events['loadingdone'] is undefined
PASS document.fontloader.checkFont('10px TestFont1, TestFont2') is true
PASS events['loadingdone'] is 1
PASS successfullyParsed is true

TEST COMPLETE
